% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialEpisim.data.R
\name{spatialEpisim.data_example}
\alias{spatialEpisim.data_example}
\title{Preaggregated WorldPop data for spatialEpisim}
\usage{
spatialEpisim.data_example(path = NULL)
}
\arguments{
\item{path}{The `path` argument passed to [system.file], which is the path
underneath the package extdata folder to a binary file. See that function's
documentation for `...` to better understand this function.}
}
\value{
An absolute path to the file(s).
}
\description{
Preaggregated WorldPop data for spatialEpisim
}
\examples{
terra::rast(spatialEpisim.data_example("COD/COD.grd"))

readxl::read_xlsx(spatialEpisim.data_example("shortlistseeddata.xlsx"))
}
